Bugfix/testing

+++ API/DTO/DailyNutritionsResponse.cs
@@ -3,7 +3,7 @@
 public class DailyNutritionsResponse
 {
     public int DailyCalories { get; set; }
-    public double DailyProtein { get; set; }
-    public double DailyFat { get; set; }
-    public double DailyCarbohydrates { get; set; }
+    public decimal DailyProtein { get; set; }
+    public decimal DailyFat { get; set; }
+    public decimal DailyCarbohydrates { get; set; }
 }
