+++ API/DTO/ProductRecord/ProductRecordRequest.cs
@@ -3,7 +3,7 @@
 public class ProductRecordRequest
 {
     public int ProductNutritionId { get; set; }
-    public double ConsumedGrams { get; set; }
+    public decimal ConsumedGrams { get; set; }
 
     public DateTime? Date { get; set; }
 }
